Handle scoping in a much better way.
Should probably explain this a little bit:
Neko and Brat have different scoping rules. In Neko, inner functions are
given copies of variables from outer scopes, rather than references. In
Brat, all variables in enclosing scopes are available and mutable.
Before, this was implemented by simply making all variables global. This
was a terrible, terrible thing, but it worked.
Almost.
I ran across a case today where this simply did _not_ work: recursive
function calls. Variables that should be local to each call are actually
global, creating a big problem. This forced me to redo how scoping of
variables is handled.
Now there is an explicit environment array that is created and can be
accessed in inner scopes. Right now, this is kind of inefficient, but at
least the behavior is correct.