Fix how arrays are made to accomodate larger ones
because of Neko's limitation on stack size and the way it apparently
copies array entries onto the stack. Now array elements are assigned
directly to an array instead of using the Neko $array() method.
Unfortunately, this meant some reworking of the parser, because assigning
to temp variables and then to the array does not work. Now, instead of
assigning results to a temp var, results meant for array elements are
assigned directly to the array.