[Parser] scope, method invocation, and escaping
updated for generating Lua. Since Lua scope is the same as Brat's, we
can tear out all that old code for scanning variables and manipulating
scope. The parser-extension also contains code for generating code for
method calls and such, which is updated to generate Lua. Lua has a
different set of keywords and is more strict about variable names, so
escaping functions have also been updated.