[Parser] Add ability to pass in result variable.
This is to allow nodes to reuse variables when possible. For example,
when populating an array all the elements can use the same temporary
variable before being stored into the array. For example, if the array
has 100 elements, we don't want 100 temp variables used. It is very easy
to run out of local stack space that way.