[Core] Arrays keep track of their length now
because you can't trust Lua to uphold Brat
semantics and we don't want to waste time filling
arrays will object.__null.

In other words, this was not working:
a = []
a[1000] = "jello world"

because Lua thinks that is just a regular table.
Now it works, but we have to carefully update array._length
and check for nils before returning anything